discord-bot-jurandir/pyproject.toml
[project]
name = "discord-bot-jurandir"
version = "1.0.0"
description = "Discord bot for managing roles and partners"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = []

